requirements.txt
aiohttp>=3.9
fastapi==0.109.2
uvicorn==0.27.1
line-bot-sdk==2.4.3
python-dotenv==1.0.1
requests>=2.28.0
python-multipart>=0.0.6
pytest>=7.0.0